add conf.py check in quickstart.sh

+++ quick_start.sh
@@ -4,6 +4,8 @@ admin=admin
 paswd=1234qwer
 email=marov@krista.ru
 
+[ -s conf.py ] || { echo -e "Create 'conf.py' from 'conf.py.template'!"; exit 1; }
+
 echo -e "Install libs."
 sudo apt-get install -y libpq-dev python-dev
